requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
flask==3.1.3
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
python-dotenv==1.2.2
requests==2.33.0
urllib3==2.6.3
waitress==3.0.2
werkzeug==3.1.7
